refactor(store-core): extract git result handlers into handlers/git.ts (audit P2-3)
First slice of the P2-3 handlers/index.ts split. Move the six git-result
handlers (diff_result / git_status_result / git_branches_result /
git_stage_result / git_unstage_result / git_commit_result), their payload
interfaces, the per-element type guards (isGitFileStatus / isGitBranch /
isDiffFile / isDiffHunk / isDiffHunkLine), the valid-status sets, and the
validateGitElements helper into a new handlers/git.ts.

Pure move, no logic change: the extracted block is byte-identical to the
removed lines. The barrel (handlers/index.ts) re-exports via
`export * from './git'`, so the public surface is unchanged and every
consumer (top-level store-core index, app, dashboard) is unaffected. The
five git element types are now imported only by git.ts and dropped from the
barrel's import list.

handlers/index.ts: 5689 -> 5399 lines.
